Further refactoring
Split definitions into their own files
