+++ src/crptr/synthetic_populations/corruptor_definitions/base.py
@@ -5,8 +5,6 @@ import crptr.position_functions as position_functions
 
 
 class Corruptors:
-
-
     def __init__(self, lookupFilesDir, encoding = 'UTF-8'):
         self.encoding = encoding
         self.lookupFilesDir = lookupFilesDir
@@ -147,140 +145,3 @@ class Corruptors:
         self.blankRecord = corrupt_record.CorruptClearRecord(
             clear_val=' '
         )
-
-
-
-class BirthCorruptors(Corruptors):
-
-    def __init__(self, labels, lookupFilesDir, encoding = 'UTF-8'):
-        self.columnLabels = labels
-        Corruptors.__init__(self, lookupFilesDir, encoding)
-
-    def setup(self):
-
-        Corruptors.setup(self)
-
-        # =====================================================================
-        # Record level
-        # =====================================================================
-        self.dayMonthSwapMarriage = corrupt_record.CorruptSwapAttributes(
-            attr1='day of parents\' marriage',
-            attr2='month of parents\' marriage',
-            attr_name_list=self.columnLabels
-        )
-
-        self.childNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='child\'s forname(s)',
-            attr2='child\'s surname',
-            attr_name_list=self.columnLabels
-        )
-
-        self.fatherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='father\'s forename',
-            attr2='father\'s surname',
-            attr_name_list=self.columnLabels
-        )
-
-        self.motherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='mother\'s forename',
-            attr2='mother\'s maiden surname',
-            attr_name_list=self.columnLabels
-        )
-
-class DeathCorruptors(Corruptors):
-
-    def __init__(self, labels, lookupFilesDir, encoding = 'UTF-8'):
-        self.columnLabels = labels
-        Corruptors.__init__(self, lookupFilesDir, encoding)
-
-
-    def setup(self):
-
-        Corruptors.setup(self)
-
-        # =====================================================================
-        # Attribute level
-        # =====================================================================
-
-
-        # =====================================================================
-        # Record level
-        # =====================================================================
-        self.dayMonthSwapDeath = corrupt_record.CorruptSwapAttributes(
-            attr1='day',
-            attr2='month',
-            attr_name_list=self.columnLabels
-        )
-
-        self.deceasedNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='forename(s) of deceased',
-            attr2='surname of deceased',
-            attr_name_list=self.columnLabels
-        )
-
-        self.fatherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='father\'s forename',
-            attr2='father\'s surname',
-            attr_name_list=self.columnLabels
-        )
-
-        self.motherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='mother\'s forename',
-            attr2='mother\'s maiden surname',
-            attr_name_list=self.columnLabels
-        )
-
-
-class MarriageCorruptors(Corruptors):
-
-    def __init__(self, labels, lookupFilesDir, encoding = 'UTF-8'):
-        self.columnLabels = labels
-        Corruptors.__init__(self, lookupFilesDir, encoding)
-
-
-    def setup(self):
-
-        Corruptors.setup(self)
-
-        self.dayMonthSwapDeath = corrupt_record.CorruptSwapAttributes(
-            attr1='day',
-            attr2='month',
-            attr_name_list=self.columnLabels
-        )
-
-        self.groomNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='forename of groom',
-            attr2='surname of groom',
-            attr_name_list=self.columnLabels
-        )
-
-        self.brideNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='forename of bride',
-            attr2='surname of bride',
-            attr_name_list=self.columnLabels
-        )
-
-        self.groomFatherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='groom\'s father\'s forename',
-            attr2='groom\'s father\'s surname',
-            attr_name_list=self.columnLabels
-        )
-
-        self.groomMotherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='groom\'s mother\'s forename',
-            attr2='groom\'s mother\'s maiden surname',
-            attr_name_list=self.columnLabels
-        )
-
-        self.brideFatherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='bride\'s father\'s forename',
-            attr2='bride\'s father\'s surname',
-            attr_name_list=self.columnLabels
-        )
-
-        self.brideMotherNameSwap = corrupt_record.CorruptSwapAttributes(
-            attr1='bride\'s mother\'s forename',
-            attr2='bride\'s mother\'s maiden surname',
-            attr_name_list=self.columnLabels
-        )
-
